requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.4
django-crispy-forms==2.4
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
